requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
astroid==2.4.2
attrs==19.3.0
Automat==20.2.0
boto3==1.14.15
botocore==1.17.15
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
cloudpickle==0.6.1
colorama==0.4.3
coloredlogs==10.0
constantly==15.1.0
contextlib2==0.6.0
cycler==0.10.0
distlib==0.3.0
distro==1.4.0
docutils==0.15.2
elasticsearch==7.8.0
future==0.17.1
fuzzywuzzy==0.18.0
gast==0.3.3
gevent==1.5.0
greenlet==0.4.16
grpcio==1.30.0
h5py==2.10.0
html5lib==1.0.1
humanfriendly==8.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.7.0
incremental==17.5.0
ipaddr==2.2.0
isort==4.3.21
jedi==0.17.1
jmespath==0.10.0
jsonschema==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
klein==17.10.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Markdown==3.2.2
matplotlib==2.2.5
mccabe==0.6.1
mock==4.0.2
msgpack==0.6.2
neovim==0.3.1
numpy==1.19.0
packaging==18.0
parso==0.7.0
pep517==0.8.2
progress==1.5
protobuf==3.12.2
PyHamcrest==2.0.2
pylint==2.5.3
pynvim==0.4.1
pyparsing==2.4.6
pyspellchecker==0.5.4
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytoml==0.1.21
pytz==2020.1
rasa-nlu==0.15.1
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.15.100
s3transfer==0.3.3
scikit-learn==0.20.4
scipy==1.5.0
simplejson==3.17.0
six==1.14.0
sklearn-crfsuite==0.3.6
tabulate==0.8.7
tensorboard==1.13.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.47.0
Twisted==20.3.0
typed-ast==1.4.1
typing==3.7.4.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
